requirements.txt
aiogram==3.4.1
fastapi==0.111.0
uvicorn[standard]
asyncpg
SQLAlchemy
python-dotenv
aiohttp